fixing git comments
